requirements.txt
langchain-azure-ai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
notion-client>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
firebase-admin>=6.2.0